statUpdate re-org
also renamed ChampStatsModule to Champion
